Moved around the ArrayWrapper [skip hazardlib]
